% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBRW.R
\name{bgrw.average}
\alias{bgrw.average}
\title{Bernoulli Growth Random Walk Average
Replicate the BGRW for different p values}
\usage{
bgrw.average(iterations = 100, steps = 100, p_values = c(0.25, 0.5, 0.75, 1))
}
\arguments{
\item{iterations}{integer number total number of iterations}

\item{steps}{integer number total number of steps}

\item{p_values}{vector of probabilities}
}
\value{
the average distance and average degree profile after \code{steps}
}
\description{
Bernoulli Growth Random Walk Average
Replicate the BGRW for different p values
}
