% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBRW.R
\name{bgrw}
\alias{bgrw}
\title{Bernoulli Growth Random Walk
Simulate an instance with a given number of steps and a probability p}
\usage{
bgrw(steps = 100, prob = 1)
}
\arguments{
\item{steps}{integer number total number of steps}

\item{prob}{probability of add a vertex at each step}
}
\value{
A list with the tree, walker position and distance after \code{steps}
}
\description{
Bernoulli Growth Random Walk
Simulate an instance with a given number of steps and a probability p
}
