% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBRW.R
\name{tbrw.with_vertex_seq}
\alias{tbrw.with_vertex_seq}
\title{Simulate a TBRW instance with a given number of
steps and a prescribed vertex sequence}
\usage{
tbrw.with_vertex_seq(steps = 100, vertex_seq = rep(1, 100))
}
\arguments{
\item{steps}{integer number total number of steps}

\item{vertex_seq}{a vector of the number of vertices to be add at each step}
}
\value{
A list with the tree, walker position, total number of vertices
and distance after \code{steps}
}
\description{
Simulate a TBRW instance with a given number of
steps and a prescribed vertex sequence
}
